requirements.txt

mysql-python
django==1.4.3

# For Tasty Pie:
mimeparse==0.1.3
python-dateutil==1.5
django-tastypie==0.9.11
